feat: add pwrsave script to network module
This changes allow to customize behavior of network modules power save
for wifi.

The used method spits out an error if the device has no wifi built in,
which confuses the user.

To prevent that I implemented a simple solution that the user only get a
warning if no wifi device present.

Also added functionality to disable the whole mechanism.
You can choose a standard rc.local entry or
a systemd service to manipulate behavior on reboots
(Enable/Disable the service)

Signed-off-by: Stephan Wendel <me@stephanwe.de>