Add support for resizing BTRFS based images
The assumption that all SBC images use ext[234] seems a bit overly
simplistic. This adds a simple check to make sure we are actually
dealing with an extN filesystem before running fsck against it.

Next it adds an attempt to handle BTRFS file systems. This is a bit
challenging since BTRFS needs to be online in order to be resized
(how odd is that?) and it seemed that mounting the filesystem at times
failed for strange reasons - which resulted in the hacky code that
attempts to unmount the loop device (even though it clearly isn't
mounted) and listing the content of the freshly created temporary mount
point. This seems pointless, but with those two lines in place I no
longer got the mount failures.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>